rm redundant `shell.nix`, enables the `default.nix` API interface
Code of the file is redundant, it simply imported the `default.nix` loosing the
interface of it in the process.

`shell.nix` was not passing any arguments into derivation, even from
`nix-shell --argstr compiler ghc8101`

even with: `{...}`. The only way passing interface seems to be to explicitly
replicate it, since the `shell.nix` code does nothing - it was removed.

Now `nix-shell` loads `default.nix` directly, so `nix-shell` and `nix-build`
have the same interface.